% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_adult.R
\docType{data}
\name{data_adult}
\alias{data_adult}
\title{Predict whether income exceeds $50K/yr based on census data}
\format{A data.frame with 48842 rows and 13 columns}
\source{
Ron Kohavi, "Scaling Up the Accuracy of Naive-Bayes Classifiers: a Decision-Tree Hybrid", Proceedings
of the Second International Conference on Knowledge Discovery and Data Mining, 1996

Dua, D. and Graff, C. (2019). UCI Machine Learning Repository \url{http://archive.ics.uci.edu/ml}. Irvine, CA:
University of California, School of Information and Computer Science.

\url{https://archive.ics.uci.edu/ml/datasets/adult}
}
\usage{
data_adult
}
\description{
This is the \code{Adult} or \code{Census Income} dataset from the UCI Machine Learning Repository.
}
\keyword{datasets}
