% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapFlex.R
\name{shapFlex}
\alias{shapFlex}
\title{Compute symmetric or asymmetric stochastic feature-level Shapley values}
\usage{
shapFlex(explain, reference = NULL, model, predict_function,
  target_features = NULL, causal = NULL, causal_weights = NULL,
  sample_size = 60, use_future = FALSE)
}
\arguments{
\item{explain}{TBD}

\item{reference}{TBD}

\item{model}{TBD}

\item{target_features}{Optional. A character vector of feature names in \code{data} for which Shapley values will be produced. For high-dimensional models, selecting a subset
of interesting features may dramatically speed up computation time. The default value is to return feature effects for all features across all input datasets.}

\item{causal}{TBD}

\item{causal_weights}{TBD}

\item{sample_size}{The number of randomly selected comparison rows or instances used to compute the feature-level Shapley value.}

\item{use_future}{Boolean. If \code{TRUE}, the \code{future} package is used to calculate Shapley values in parallel.}

\item{predict_functions}{TBD}
}
\value{
A data.frame of the feature-level Shapley values for the target instance and selected features.
}
\description{
This function uses user-defined trained models and prediction functions to compute approximate Shapley values for
single models. Shapley values can be calculated for a subset of model features which reduces the
typically expensive computation of approximate Shapley values in high-dimensional models.
}
